Don't rely on /etc/hosts to get the internal API VIP
The internal API VIP entry in /etc/hosts is not avialable in newton, but
we can get the value from hiera. We still rely on /etc/hosts to get the
rest of the IPs and fqdns for the nodes, so in releases newer than
newton we still need to filter out the internal API VIP FQDN from the
output.