Make resource agent creation generic
This now grabs the network list and creates the resource
agents from it. This requires a bit more complex query, where it
will get all the VIP values and flatten the list. But with this
addition, the storage network VIP resource agent is now created
(since it's in the list).