Skip ctlplane network and fallback IPs
When a certain network is not configured for a certain role, the
default behavior is to fallback to the ctlplane address. We explicitly
skip these addresses, as well as the whole ctlplane network.

By skipping the ctlplane network we can now provide a strict policy
for the IPSEC tunnels without loosing connectivity.