Move host check outside of resource-agent.yml
We only want to create resource agents in one node. The way it was
done in the legacy workflow was to check for a specific host with
a specific variable. This was moved outside of the tasks that
actually create the resource agent, with the purpose of making
this check the responsibility of the caller.