Move firewall setup to main.yml
I left it in the legacy setup, expecting to do the same in the newer
one. But I didn't. So this turned out problematic.

I moved it tot he main.yml file since it's an overall feature of the
role, and this way it's also explicit that it applies to both the newer
and the legacy setups.