WORKAROUND: Use include_tasks instead of calling role
It was referencing the old playbook's role name. And for some
reason, the legacy lint job installs the job in a directory called
"workspace", which is not ideal since it's what it uses as the role
name. So instead of using the actual role name, we call the task
directly.

Note that it also fixes a trailing whitespace from the meta/mail.yml
file

Change-Id: I89a42b72be08a1171e2c1dc7b7c0a14caad8d634