Fix Zuul v3 file name
Zuul v3 currently only parses .yaml files, rename the file.

remove the project name stanza, it is not needed and will harm with
project renames. The current automated changes will not catch this, as
this basically is a new name.

Also, remove .zuul.yaml - we don't need both files. And the noop
template is just wrong for this project since it has jobs.

Change-Id: I112711a28d9635cb9eaa6f788e1747b20596701b